+++ shared/src/MapsCoreSharedModule.cpp
@@ -10,4 +10,4 @@
 
 #include "MapsCoreSharedModule.h"
 
-std::string MapsCoreSharedModule::version() { return "2.0.0"; }
+std::string MapsCoreSharedModule::version() { return "2.0.1"; }